+++ Samples~/Prototype/Bootstraps/PrototypeBootstrap.cs
@@ -1,35 +1,17 @@
 ﻿using EcsRx.Zenject;
 using EcsRx.Infrastructure.Extensions;
 using UnityEngine;
-using Inter.Installer;
-using Inter.Modules.ToolModule;
-using Inter.Prototype.Installer;
+using InterVR.Unity.SDK.InterFramework.Installer;
+using InterVR.Unity.SDK.InterFramework.Prototype.Installer;
 
-namespace Inter.Prototype
+namespace InterVR.Unity.SDK.InterFramework.Prototype
 {
-    public class PrototypeBootstrap : EcsRxApplicationBehaviour
+    public class PrototypeBootstrap : InterFrameworkApplicationBehaviour
     {
-        protected override void BindSystems()
-        {
-            base.BindSystems();
-        }
-
-        protected override void LoadModules()
-        {
-            base.LoadModules();
-
-            Container.LoadModule<ToolModuleSetupDependency>();
-        }
-
-        protected override void LoadPlugins()
-        {
-            base.LoadPlugins();
-        }
-
         protected override void ApplicationStarted()
         {
             var settings = Container.Resolve<PrototypeInstaller.Settings>();
-            var interSettings = Container.Resolve<InterInstaller.Settings>();
+            var interSettings = Container.Resolve<InterFrameworkInstaller.Settings>();
             Debug.Log($"settings.Name is {settings.Name} in {interSettings.Name}");
         }