+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this package will be documented in this file.
 
+## [0.3.0] - 2020-08-13
+
+### Changed
+
+- Change entire namespace prefix to 'InterVR.Unity.SDK.InterFramework'
+
+### Added
+
+- InterFrameworkApplicationBehaviour derrived from EcsRxApplicationBehaviour
+
 ## [0.2.0] - 2020-08-13
 
 ### Changed